20230515.02
LE fond d'écran change uniquement pour les installations SonXPlus

+++ installWindows.ps1
@@ -416,11 +416,13 @@ function CleanDesktop {
     }
     while (!(Get-ItemProperty $DesktopPath"\Programmes\TeamViewer.lnk" -erroraction 'silentlycontinue'))
 
-    #Changement du fond d'écran
-    
-    $wallpaperPath = "$env:USERPROFILE\AppData\Roaming\Microsoft\Windows\Themes"
-    Remove-Item -Path "$wallpaperPath\TranscodedWallpaper"
-    Copy-Item -Path .\extension\wallpaper.jpg -Destination "$wallpaperPath\TranscodedWallpaper"
+    #Changement du fond d'écran si installation est SonXPlus
+    if($installationType -eq 2)
+    {
+        $wallpaperPath = "$env:USERPROFILE\AppData\Roaming\Microsoft\Windows\Themes"
+        Remove-Item -Path "$wallpaperPath\TranscodedWallpaper"
+        Copy-Item -Path .\extension\wallpaper.jpg -Destination "$wallpaperPath\TranscodedWallpaper"
+    }
 
 }
 # ------------------------------------------------------------------------------------------------------------------ #
